+++ docs/news.rst
@@ -3,6 +3,18 @@
 Release notes
 =============
 
+Scrapy VERSION (unreleased)
+---------------------------
+
+Backward-incompatible changes
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+-   If you set the :setting:`TWISTED_REACTOR` setting to a :ref:`non-asyncio
+    value <disable-asyncio>` at the :ref:`spider level <spider-settings>`, you
+    may now need to set the :setting:`FORCE_CRAWLER_PROCESS` setting to
+    ``True`` when running Scrapy via :ref:`its command-line tool
+    <topics-commands-crawlerprocess>` to avoid a reactor mismatch exception.
+
 .. _release-2.13.1:
 
 Scrapy 2.13.1 (2025-05-28)